chore: move skip to ext
